errors

+++ billing/invoice/models/invoice.py
@@ -7,10 +7,10 @@ from .discount import Discount
 
 class Invoices(BaseModel):
 
-    total       = models.BigIntegerField('Total factura: ')
+    total       = models.FloatField('Total factura: ')
     client      = models.ForeignKey(Client,  on_delete=models.CASCADE, blank=True, null=True, related_name='invoicesclient')
     seller      = models.ForeignKey(Seller,on_delete=models.CASCADE, related_name='invoicesseller')
-    isv         = models.BigIntegerField('ISV 15%')
+    isv         = models.FloatField('ISV 15%')
     discount    = models.ForeignKey(Discount,on_delete=models.CASCADE, blank=True, null=True, related_name='invoicesdicount')
     status      = models.BooleanField('Estado', default=True)
     def __str__(self) :
